pyproject.toml
[project]
name = "marina-vista-reception"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gradio>=5.49.1",
    "openai>=2.6.1",
    "python-dotenv>=1.2.1",
]
